chore: extracted out ./src/utils/curriculum/openapi + tests
