+++ src/utils/curriculum/openapi.test.ts
@@ -1,0 +1,1719 @@
+import openApiRequest from "./openapi";
+
+const fetch = jest.spyOn(global, "fetch") as jest.Mock;
+
+describe("openapi", () => {
+  it("units", async () => {
+    const body = JSON.stringify([
+      {
+        year: 7,
+        units: [
+          {
+            unitTitle: "Place value",
+            unitSlug: "place-value",
+            order: 1,
+          },
+          {
+            unitTitle:
+              "Properties of number: factors, multiples, squares and cubes",
+            unitSlug:
+              "properties-of-number-factors-multiples-squares-and-cubes",
+            order: 2,
+          },
+        ],
+      },
+      {
+        year: 8,
+        units: [
+          {
+            unitTitle: "Place value",
+            unitSlug: "place-value",
+            order: 1,
+          },
+          {
+            unitTitle:
+              "Properties of number: factors, multiples, squares and cubes",
+            unitSlug:
+              "properties-of-number-factors-multiples-squares-and-cubes",
+            order: 2,
+          },
+        ],
+      },
+    ]);
+    fetch.mockResolvedValue(new Response(body, { status: 200 }));
+    const output = await openApiRequest("test", {
+      subjectSlug: "english",
+      phaseSlug: "primary",
+      ks4OptionSlug: null,
+    });
+    expect(output).toEqual({
+      units: [
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+      ],
+    });
+  });
+
+  it("units + subjects", async () => {
+    const body = JSON.stringify([
+      {
+        year: 7,
+        subjects: [
+          {
+            subject: "test1",
+            units: [
+              {
+                unitTitle: "Place value",
+                unitSlug: "place-value",
+                order: 1,
+              },
+              {
+                unitTitle:
+                  "Properties of number: factors, multiples, squares and cubes",
+                unitSlug:
+                  "properties-of-number-factors-multiples-squares-and-cubes",
+                order: 2,
+              },
+            ],
+          },
+          {
+            subject: "test2",
+            units: [
+              {
+                unitTitle: "Place value",
+                unitSlug: "place-value",
+                order: 1,
+              },
+              {
+                unitTitle:
+                  "Properties of number: factors, multiples, squares and cubes",
+                unitSlug:
+                  "properties-of-number-factors-multiples-squares-and-cubes",
+                order: 2,
+              },
+            ],
+          },
+        ],
+      },
+      {
+        year: 8,
+        subjects: [
+          {
+            subject: "test1",
+            units: [
+              {
+                unitTitle: "Place value",
+                unitSlug: "place-value",
+                order: 1,
+              },
+              {
+                unitTitle:
+                  "Properties of number: factors, multiples, squares and cubes",
+                unitSlug:
+                  "properties-of-number-factors-multiples-squares-and-cubes",
+                order: 2,
+              },
+            ],
+          },
+          {
+            subject: "test2",
+            units: [
+              {
+                unitTitle: "Place value",
+                unitSlug: "place-value",
+                order: 1,
+              },
+              {
+                unitTitle:
+                  "Properties of number: factors, multiples, squares and cubes",
+                unitSlug:
+                  "properties-of-number-factors-multiples-squares-and-cubes",
+                order: 2,
+              },
+            ],
+          },
+        ],
+      },
+    ]);
+    fetch.mockResolvedValue(new Response(body, { status: 200 }));
+    const output = await openApiRequest("test", {
+      subjectSlug: "english",
+      phaseSlug: "primary",
+      ks4OptionSlug: null,
+    });
+    // console.log("output=", JSON.stringify(output, null, 2))
+    expect(output).toEqual({
+      units: [
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: null,
+          tier_slug: null,
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+      ],
+    });
+  });
+
+  it("units + tiers", async () => {
+    const body = JSON.stringify([
+      {
+        year: 7,
+        tiers: [
+          {
+            tier: "test1",
+            units: [
+              {
+                unitTitle: "Place value",
+                unitSlug: "place-value",
+                order: 1,
+              },
+              {
+                unitTitle:
+                  "Properties of number: factors, multiples, squares and cubes",
+                unitSlug:
+                  "properties-of-number-factors-multiples-squares-and-cubes",
+                order: 2,
+              },
+            ],
+          },
+          {
+            tier: "test2",
+            units: [
+              {
+                unitTitle: "Place value",
+                unitSlug: "place-value",
+                order: 1,
+              },
+              {
+                unitTitle:
+                  "Properties of number: factors, multiples, squares and cubes",
+                unitSlug:
+                  "properties-of-number-factors-multiples-squares-and-cubes",
+                order: 2,
+              },
+            ],
+          },
+        ],
+      },
+      {
+        year: 8,
+        tiers: [
+          {
+            tier: "test1",
+            units: [
+              {
+                unitTitle: "Place value",
+                unitSlug: "place-value",
+                order: 1,
+              },
+              {
+                unitTitle:
+                  "Properties of number: factors, multiples, squares and cubes",
+                unitSlug:
+                  "properties-of-number-factors-multiples-squares-and-cubes",
+                order: 2,
+              },
+            ],
+          },
+          {
+            tier: "test2",
+            units: [
+              {
+                unitTitle: "Place value",
+                unitSlug: "place-value",
+                order: 1,
+              },
+              {
+                unitTitle:
+                  "Properties of number: factors, multiples, squares and cubes",
+                unitSlug:
+                  "properties-of-number-factors-multiples-squares-and-cubes",
+                order: 2,
+              },
+            ],
+          },
+        ],
+      },
+    ]);
+    fetch.mockResolvedValue(new Response(body, { status: 200 }));
+    const output = await openApiRequest("test", {
+      subjectSlug: "english",
+      phaseSlug: "primary",
+      ks4OptionSlug: null,
+    });
+    // console.log("output=", JSON.stringify(output, null, 2))
+    expect(output).toEqual({
+      units: [
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "english",
+          subject_slug: "english",
+          subject_parent: null,
+          subject_parent_slug: null,
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+      ],
+    });
+  });
+
+  it("units subjects + tiers", async () => {
+    const body = JSON.stringify([
+      {
+        year: 7,
+        subjects: [
+          {
+            subject: "test1",
+            tiers: [
+              {
+                tier: "test1",
+                units: [
+                  {
+                    unitTitle: "Place value",
+                    unitSlug: "place-value",
+                    order: 1,
+                  },
+                  {
+                    unitTitle:
+                      "Properties of number: factors, multiples, squares and cubes",
+                    unitSlug:
+                      "properties-of-number-factors-multiples-squares-and-cubes",
+                    order: 2,
+                  },
+                ],
+              },
+              {
+                tier: "test2",
+                units: [
+                  {
+                    unitTitle: "Place value",
+                    unitSlug: "place-value",
+                    order: 1,
+                  },
+                  {
+                    unitTitle:
+                      "Properties of number: factors, multiples, squares and cubes",
+                    unitSlug:
+                      "properties-of-number-factors-multiples-squares-and-cubes",
+                    order: 2,
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            subject: "test2",
+            tiers: [
+              {
+                tier: "test1",
+                units: [
+                  {
+                    unitTitle: "Place value",
+                    unitSlug: "place-value",
+                    order: 1,
+                  },
+                  {
+                    unitTitle:
+                      "Properties of number: factors, multiples, squares and cubes",
+                    unitSlug:
+                      "properties-of-number-factors-multiples-squares-and-cubes",
+                    order: 2,
+                  },
+                ],
+              },
+              {
+                tier: "test2",
+                units: [
+                  {
+                    unitTitle: "Place value",
+                    unitSlug: "place-value",
+                    order: 1,
+                  },
+                  {
+                    unitTitle:
+                      "Properties of number: factors, multiples, squares and cubes",
+                    unitSlug:
+                      "properties-of-number-factors-multiples-squares-and-cubes",
+                    order: 2,
+                  },
+                ],
+              },
+            ],
+          },
+        ],
+      },
+      {
+        year: 8,
+        subjects: [
+          {
+            subject: "test1",
+            tiers: [
+              {
+                tier: "test1",
+                units: [
+                  {
+                    unitTitle: "Place value",
+                    unitSlug: "place-value",
+                    order: 1,
+                  },
+                  {
+                    unitTitle:
+                      "Properties of number: factors, multiples, squares and cubes",
+                    unitSlug:
+                      "properties-of-number-factors-multiples-squares-and-cubes",
+                    order: 2,
+                  },
+                ],
+              },
+              {
+                tier: "test2",
+                units: [
+                  {
+                    unitTitle: "Place value",
+                    unitSlug: "place-value",
+                    order: 1,
+                  },
+                  {
+                    unitTitle:
+                      "Properties of number: factors, multiples, squares and cubes",
+                    unitSlug:
+                      "properties-of-number-factors-multiples-squares-and-cubes",
+                    order: 2,
+                  },
+                ],
+              },
+            ],
+          },
+          {
+            subject: "test2",
+            tiers: [
+              {
+                tier: "test1",
+                units: [
+                  {
+                    unitTitle: "Place value",
+                    unitSlug: "place-value",
+                    order: 1,
+                  },
+                  {
+                    unitTitle:
+                      "Properties of number: factors, multiples, squares and cubes",
+                    unitSlug:
+                      "properties-of-number-factors-multiples-squares-and-cubes",
+                    order: 2,
+                  },
+                ],
+              },
+              {
+                tier: "test2",
+                units: [
+                  {
+                    unitTitle: "Place value",
+                    unitSlug: "place-value",
+                    order: 1,
+                  },
+                  {
+                    unitTitle:
+                      "Properties of number: factors, multiples, squares and cubes",
+                    unitSlug:
+                      "properties-of-number-factors-multiples-squares-and-cubes",
+                    order: 2,
+                  },
+                ],
+              },
+            ],
+          },
+        ],
+      },
+    ]);
+    fetch.mockResolvedValue(new Response(body, { status: 200 }));
+    const output = await openApiRequest("test", {
+      subjectSlug: "english",
+      phaseSlug: "primary",
+      ks4OptionSlug: null,
+    });
+    expect(output).toEqual({
+      units: [
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "7",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test1",
+          subject_slug: "test1",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test1",
+          tier_slug: "test1",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 1,
+          slug: "place-value",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Place value",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+        {
+          connection_prior_unit_description: null,
+          connection_future_unit_description: null,
+          connection_future_unit_title: null,
+          connection_prior_unit_title: null,
+          domain: null,
+          domain_id: null,
+          examboard: null,
+          examboard_slug: null,
+          planned_number_of_lessons: 0,
+          phase: "primary",
+          phase_slug: "primary",
+          keystage_slug: "",
+          lessons: [],
+          order: 2,
+          slug: "properties-of-number-factors-multiples-squares-and-cubes",
+          subject: "test2",
+          subject_slug: "test2",
+          subject_parent: "english",
+          subject_parent_slug: "english",
+          tier: "test2",
+          tier_slug: "test2",
+          pathway: null,
+          pathway_slug: null,
+          tags: [],
+          subjectcategories: [],
+          threads: [],
+          title: "Properties of number: factors, multiples, squares and cubes",
+          description: "",
+          why_this_why_now: "",
+          cycle: "2",
+          features: null,
+          unit_options: [],
+          year: "8",
+          state: "published",
+        },
+      ],
+    });
+  });
+});
